requirements.txt
tkinter==8.6
Pillow==9.5.0
ultralytics==8.1.3